refactor: format dto packages to team standards
