+++ packages/jaree-server-repo/src/main/java/org/jaree/api/application/output/ApplicationOutputDTO.java
@@ -1,4 +1,4 @@
-package org.jaree.api.application.dto;
+package org.jaree.api.application.output;
 
 import java.time.LocalDateTime;
 import java.util.List;
