+++ packages/jaree-server-repo/src/main/java/org/jaree/api/application/output/ApplicationVersionCommitMessageDTO.java
@@ -1,4 +1,4 @@
-package org.jaree.api.application.dto;
+package org.jaree.api.application.output;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -15,27 +15,23 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ApplicationVersionSimpleDTO {
+public class ApplicationVersionCommitMessageDTO {
     private String id;
     private String commitMessage;
     private LocalDateTime createdAt;
     private List<String> ancestorIds;
-    private List<ApplicationAnswerSimpleDTO> answers;
 
-    public static ApplicationVersionSimpleDTO of(ApplicationVersion applicationVersion) {
+    public static ApplicationVersionCommitMessageDTO of(ApplicationVersion applicationVersion) {
         if (applicationVersion == null) {
             return null;
         }
 
-        return ApplicationVersionSimpleDTO.builder()
+        return ApplicationVersionCommitMessageDTO.builder()
             .id(applicationVersion.getId())
             .commitMessage(applicationVersion.getCommitMessage())
             .createdAt(applicationVersion.getCreatedAt())
             .ancestorIds(applicationVersion.getAncestors().stream()
                 .map(ApplicationVersion::getId)
-                .collect(Collectors.toList()))
-            .answers(applicationVersion.getAnswers().stream()
-                .map(ApplicationAnswerSimpleDTO::of)
                 .collect(Collectors.toList()))
             .build();
     }
