+++ packages/jaree-server-repo/src/main/java/org/jaree/api/application/service/ApplicationService.java
@@ -5,14 +5,14 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.jaree.api.application.dto.ApplicationAnswerCreationInputDTO;
-import org.jaree.api.application.dto.ApplicationCreationInputDTO;
-import org.jaree.api.application.dto.ApplicationOutputDTO;
+import org.jaree.api.application.input.ApplicationAnswerCreationInputDTO;
+import org.jaree.api.application.input.ApplicationCreationInputDTO;
+import org.jaree.api.application.output.ApplicationOutputDTO;
 import org.jaree.api.application.dto.ApplicationQuestionSimpleDTO;
-import org.jaree.api.application.dto.ApplicationUpdateInputDTO;
-import org.jaree.api.application.dto.ApplicationVersionCommitMessageDTO;
-import org.jaree.api.application.dto.ApplicationVersionCreationInputDTO;
-import org.jaree.api.application.dto.ApplicationVersionSimpleDTO;
+import org.jaree.api.application.input.ApplicationUpdateInputDTO;
+import org.jaree.api.application.output.ApplicationVersionCommitMessageDTO;
+import org.jaree.api.application.input.ApplicationVersionCreationInputDTO;
+import org.jaree.api.application.output.ApplicationVersionSimpleDTO;
 import org.jaree.api.application.entity.Application;
 import org.jaree.api.application.entity.ApplicationAnswer;
 import org.jaree.api.application.entity.ApplicationQuestion;
