+++ packages/jaree-server-repo/src/main/java/org/jaree/api/jobopening/output/JobOpeningOutputDTO.java
@@ -1,10 +1,11 @@
-package org.jaree.api.jobopening.dto;
+package org.jaree.api.jobopening.output;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 
-import org.jaree.api.application.dto.ApplicationQuestionOutputDTO;
+import org.jaree.api.application.output.ApplicationQuestionOutputDTO;
+import org.jaree.api.company.output.CompanyOutputDTO;
 import org.jaree.api.jobopening.entity.JobOpening;
 
 import lombok.AllArgsConstructor;
@@ -12,7 +13,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class JobOpeningWithoutCompanyOutputDTO {
+public class JobOpeningOutputDTO{
     private String id;
     private String title;
     private String description;
@@ -21,10 +22,14 @@ public class JobOpeningWithoutCompanyOutputDTO {
     private LocalDateTime startsAt;
     private LocalDateTime endsAt;
     private LocalDateTime createdAt;
+    private CompanyOutputDTO company;
     private List<ApplicationQuestionOutputDTO> questions;
 
-    public static JobOpeningWithoutCompanyOutputDTO from(JobOpening jobOpening){
+    public static JobOpeningOutputDTO from(JobOpening jobOpening){
+
         if(jobOpening == null) return null;
+
+        CompanyOutputDTO company = CompanyOutputDTO.from(jobOpening.getCompany());
 
         List<ApplicationQuestionOutputDTO> questions = jobOpening.getQuestions() == null
             ? List.of()
@@ -33,12 +38,12 @@ public class JobOpeningWithoutCompanyOutputDTO {
                 .map(ApplicationQuestionOutputDTO::from)
                 .toList();
 
-        return new JobOpeningWithoutCompanyOutputDTO(
+        return new JobOpeningOutputDTO(
             jobOpening.getId(), jobOpening.getTitle(),
             jobOpening.getDescription(), jobOpening.getContentS3Url(),
             jobOpening.getImageUrl(), jobOpening.getStartsAt(),
             jobOpening.getEndsAt(), jobOpening.getCreatedAt(),
-            questions
+            company, questions
         );
     }
 }
