+++ packages/jaree-server-repo/src/main/java/org/jaree/api/jobopening/output/JobOpeningWithoutCompanyOutputDTO.java
@@ -1,11 +1,10 @@
-package org.jaree.api.jobopening.dto;
+package org.jaree.api.jobopening.output;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 
-import org.jaree.api.application.dto.ApplicationQuestionOutputDTO;
-import org.jaree.api.company.dto.CompanyOutputDTO;
+import org.jaree.api.application.output.ApplicationQuestionOutputDTO;
 import org.jaree.api.jobopening.entity.JobOpening;
 
 import lombok.AllArgsConstructor;
@@ -13,7 +12,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class JobOpeningOutputDTO{
+public class JobOpeningWithoutCompanyOutputDTO {
     private String id;
     private String title;
     private String description;
@@ -22,14 +21,10 @@ public class JobOpeningOutputDTO{
     private LocalDateTime startsAt;
     private LocalDateTime endsAt;
     private LocalDateTime createdAt;
-    private CompanyOutputDTO company;
     private List<ApplicationQuestionOutputDTO> questions;
 
-    public static JobOpeningOutputDTO from(JobOpening jobOpening){
-
+    public static JobOpeningWithoutCompanyOutputDTO from(JobOpening jobOpening){
         if(jobOpening == null) return null;
-
-        CompanyOutputDTO company = CompanyOutputDTO.from(jobOpening.getCompany());
 
         List<ApplicationQuestionOutputDTO> questions = jobOpening.getQuestions() == null
             ? List.of()
@@ -38,12 +33,12 @@ public class JobOpeningOutputDTO{
                 .map(ApplicationQuestionOutputDTO::from)
                 .toList();
 
-        return new JobOpeningOutputDTO(
+        return new JobOpeningWithoutCompanyOutputDTO(
             jobOpening.getId(), jobOpening.getTitle(),
             jobOpening.getDescription(), jobOpening.getContentS3Url(),
             jobOpening.getImageUrl(), jobOpening.getStartsAt(),
             jobOpening.getEndsAt(), jobOpening.getCreatedAt(),
-            company, questions
+            questions
         );
     }
 }
